Added support for collection
